Revert "Added folder structure for objects"
This reverts commit f898928a61fc8f837c63dcd54d19699bb7306224.
