--- game_objects/button.py
@@ -1,34 +1,0 @@
-# TODO: add imports
-
-class Button:
-    # button class
-    def __init__(self, x, y, image, scale):
-        width = image.get_width()
-        height = image.get_height()
-        self.image = pygame.transform.scale(image, (int(width * scale), int(height * scale)))
-        self.rect = self.image.get_rect()
-        self.rect.center = (x, y)
-        self.clicked = False
-
-    def draw(self, surface):
-        action = False
-
-        # get mouse position
-        pos = pygame.mouse.get_pos()
-
-        # check mouseover and clicked conditions
-        if self.rect.collidepoint(pos):
-            if pygame.mouse.get_pressed()[0] == 1 and not self.clicked:
-                action = True
-                self.clicked = True
-
-        if pygame.mouse.get_pressed()[0] == 0:
-            self.clicked = False
-
-        # draw button
-        surface.blit(self.image, (self.rect.x, self.rect.y))
-
-        return action
-
-    def update_pos(self, new_x: int, new_y: int):
-        self.rect.center = (new_x, new_y)
