--- game_objects/ship.py
@@ -1,241 +1,0 @@
-import pygame
-import math
-import load
-
-# TODO: add imports
-# TODO: fix errors
-
-image = load.game_images()
-
-class Ship(pygame.sprite.Sprite):
-    def __init__(self, speed):
-        pygame.sprite.Sprite.__init__(self)
-        self.energy_stor = None
-        self.flip = False
-        self.max_speed = speed
-        self.speed = 0
-        self.inertia_factor = 0.02
-        self.angle = 0
-        self.moving_direction = []  # dir_vector
-        # energy things and rectangles
-        self.energy_full = 100
-        self.energy = 100
-        self.energy_max = pygame.Rect(70, 650, 100, 40)
-        self.multiple_keys = False
-        # initial credits
-        self.credits = 10
-        # storage
-        self.storage_max = 15  # HERE adjust to change maximum storage
-        self.storage = 0
-        # shooting
-        self.cooldown = 30
-        self.damage = 1
-
-        # Render ship
-        self.update_time = pygame.time.get_ticks()
-        self.index = 0
-        self.image = image['ship_0']  # default ship frame (idle)
-        self.render_state0 = image['ship_0']  # idle frame
-        self.ship_animation_frames = [image['ship_1'], image['ship_2']]  # list of moving frames for the ship
-
-        self.rect = self.image.get_rect()
-        # this sets starting position
-        self.rect.center = (SCREEN_WIDTH // 2, SCREEN_HEIGHT // 2)
-
-        # variables
-        self.moving_w = False
-        self.moving_s = False
-        self.moving_a = False
-        self.moving_d = False
-        self.shooting = False
-
-    def update(self):
-        self.action()
-        self.moving()
-
-        # print(f'Current speed: {self.speed}\n\n')
-
-        if (
-                self.moving_s or self.moving_w or self.moving_a or self.moving_d) and self.energy > 0 and not self.multiple_keys:
-            self.render_ship_animation()  # this runs ship animation logic
-            if self.speed < self.max_speed:
-                self.speed += self.max_speed * self.inertia_factor
-            else:
-                self.speed = self.max_speed
-        else:
-            self.image = image['ship_0']  # this resets ship's frame to idle if it is not moving
-            if self.speed > 0:
-                self.speed -= self.max_speed * self.inertia_factor
-                t = self.speed / math.sqrt(self.moving_direction[0] ** 2 + self.moving_direction[1] ** 2)
-                move_objects(*self.moving_direction, t)
-            else:
-                self.speed = 0
-
-        self.draw_ship()
-
-    def moving(self):
-
-        energy_consumed = 0
-        self.energy = round(self.energy, 2)
-        # movement variable is calculated (and rounded down to prevent floating position coords), and then used to move objects
-        movement_variable = round(self.speed * 0.75, 0)
-
-        if self.energy > 0 and not self.multiple_keys:
-            direction_vector = self.get_mouse_vector()
-            t = movement_variable / math.sqrt(direction_vector[0] ** 2 + direction_vector[1] ** 2)
-
-            if self.moving_w:
-                energy_consumed -= 1
-                move_objects(-direction_vector[0], -direction_vector[1], t)
-                self.moving_direction = [-direction_vector[0], -direction_vector[1]]
-
-            if self.moving_s:
-                energy_consumed -= 1
-                move_objects(direction_vector[0], direction_vector[1], t)
-                self.moving_direction = [direction_vector[0], direction_vector[1]]
-
-            if self.moving_a:
-                energy_consumed -= 1
-                move_objects(-direction_vector[1], 0, t)
-                self.moving_direction = [-direction_vector[0], 0]
-
-            if self.moving_d:
-                energy_consumed -= 1
-                move_objects(direction_vector[1], 0, t)
-                self.moving_direction = [direction_vector[1], 0]
-
-        if (self.moving_s and self.moving_w) or (self.moving_a and self.moving_d):
-            energy_consumed = 0
-            self.multiple_keys = True
-
-        else:
-            self.multiple_keys = False
-
-        # handle energy consuming (so that it won't burn 2 units if pressing W and S for example)
-        if energy_consumed <= -1:
-            self.energy -= 0.3  # HERE change the parameter to adjust energy consuming speed
-
-    def action(self):
-        # check for low or max energy
-        if self.energy <= self.energy_full * 0.3 and not self.energy <= 0:
-            draw_text('WARNING, LOW ENERGY', font_small, RED, 25, 105)
-
-        elif self.energy <= 0:
-            draw_text('NO ENERGY', font_small, RED, 25, 105)
-
-        if self.energy > self.energy_full:
-            self.energy = self.energy_full
-
-        # draw ENERGY gauge
-        draw_text('ENERGY', font_small, WHITE, 70, 10)
-        bar_length, bar_width = 48, 7
-        energy_stor = pygame.Rect(25 + bar_length, 50 + bar_width, int(94 * (self.energy / 100)), 36)
-
-        pygame.draw.rect(screen, EMPTY_BLACK, (25 + bar_length, 50 + bar_width, 94, 36))
-        pygame.draw.rect(screen, ENERGY_BLUE, energy_stor)
-        screen.blit(image['energy_bar'], (25, 50))
-
-        # inventory and money
-        self.credits = round(self.credits, 2)
-        draw_text('COIN', font_small, WHITE, 300, 10)
-        screen.blit(image['coin'], (300, 50))
-        draw_text(f'{self.credits}', font_small, WHITE, 350, 55)
-
-        if self.credits <= 0:
-            self.credits = 0
-
-        # cargo
-        self.storage = round(self.storage, 2)
-        draw_text('CARGO', font_small, WHITE, 545, 10)
-        cargo_stored = pygame.Rect(500 + bar_length, 50 + bar_width, int(94 * (self.storage / 15)), 36)
-
-        pygame.draw.rect(screen, EMPTY_BLACK, (500 + bar_length, 50 + bar_width, 94, 36))
-        pygame.draw.rect(screen, STORAGE_BROWN, cargo_stored)
-        screen.blit(image['storage_bar'], (500, 50))
-
-        if self.storage >= self.storage_max * 0.75 and not self.storage == self.storage_max:
-            draw_text(f'Reaching maximum capacity', font_small, RED, 500, 105)
-
-        elif self.storage >= self.storage_max:
-            draw_text(f'Maximum cargo capacity reached', font_small, RED, 500, 105)
-
-        if self.storage < 0:
-            self.storage = 0
-
-        if self.storage > self.storage_max:
-            self.storage = self.storage_max
-
-        if self.shooting and self.cooldown <= 0 and self.energy > 0:
-            # Spawns a laser sprite in the direction the ship is facing
-
-            spawn_dist_from_player = 40
-            dir_vector = self.get_mouse_vector()
-            t = spawn_dist_from_player / math.sqrt(dir_vector[0] ** 2 + dir_vector[1] ** 2)
-            parametric_equation_x = self.rect.centerx + (t * dir_vector[0])
-            parametric_equation_y = self.rect.centery + (t * dir_vector[1])
-
-            laser = Laser(parametric_equation_x, parametric_equation_y, dir_vector, self.angle)
-            laser_group.add(laser)
-            Player.energy -= 1
-            self.cooldown = 30
-
-        self.cooldown -= 1
-
-    def render_ship_animation(self):
-        animation_cooldown = 300  # HERE you set up animation speed
-        self.image = self.ship_animation_frames[self.index]
-
-        if pygame.time.get_ticks() - self.update_time > animation_cooldown:
-            self.index += 1
-            self.update_time = pygame.time.get_ticks()
-
-            if self.index > 1:
-                self.index = 0
-
-    def get_mouse_vector(self):
-        """Returns a reduced direction vector of the line between the cursor and player position"""
-        mouse_pos = pygame.mouse.get_pos()
-        player_pos = self.rect.center
-
-        x_component = mouse_pos[0] - player_pos[0]
-        y_component = mouse_pos[1] - player_pos[1]
-
-        if x_component != 0 and y_component != 0:
-            reduced_fraction = Fraction(x_component, y_component)
-            if (x_component < 0 and y_component < 0) or (x_component > 0 > y_component):
-                x_component = reduced_fraction.numerator * -1
-                y_component = reduced_fraction.denominator * -1
-
-        return [x_component, y_component]
-
-    def determine_direction(self):
-        """Determines the ship direction in degrees based on the cursor position and returns a rotated image"""
-        mouse_pos = pygame.mouse.get_pos()
-        adjacent = mouse_pos[0] - self.rect.centerx
-        opposite = self.rect.centery - mouse_pos[1]
-
-        if adjacent == 0 and opposite >= 0:
-            radian_angle = math.pi / 2
-        elif adjacent == 0 and opposite < 0:
-            radian_angle = (3 * math.pi) / 2
-        elif adjacent > 0 and opposite >= 0:
-            radian_angle = math.atan(opposite / adjacent)
-        elif adjacent > 0 >= opposite:
-            radian_angle = 2 * math.pi - abs(math.atan(opposite / adjacent))
-        elif adjacent < 0 and opposite <= 0:
-            radian_angle = math.pi + math.atan(opposite / adjacent)
-        elif adjacent < 0 <= opposite:
-            radian_angle = math.pi - abs(math.atan(opposite / adjacent))
-        else:
-            radian_angle = 0
-
-        self.angle = math.degrees(radian_angle)
-
-        return pygame.transform.rotate(self.image, self.angle)
-
-    def draw_ship(self):
-        self.rect.center = (screen_width//2, screen_height//2)
-        rotated_ship = self.determine_direction()
-        ship_rect = rotated_ship.get_rect(center=self.rect.center)
-
-        screen.blit(rotated_ship, ship_rect)
